refactor : LottoGenerator 에서 LottoShop 으로 이름 변경
